Minor re-factoring related to packages removing and renaming.
