+++ src/main/java/com/proxiad/task/ivanboyukliev/hangmangame/service/ApplicationConstants.java
@@ -1,4 +1,4 @@
-package com.proxiad.task.ivanboyukliev.hangmangame.util;
+package com.proxiad.task.ivanboyukliev.hangmangame.service;
 
 public final class ApplicationConstants {
