+++ src/main/java/com/proxiad/task/ivanboyukliev/hangmangame/service/InvalidGameSessionException.java
@@ -1,4 +1,4 @@
-package com.proxiad.task.ivanboyukliev.hangmangame.exception;
+package com.proxiad.task.ivanboyukliev.hangmangame.service;
 
 import jakarta.servlet.ServletException;
