+++ src/main/resources/META-INF/services/jakarta.servlet.ServletContainerInitializer
@@ -1,1 +1,1 @@
-com.proxiad.task.ivanboyukliev.hangmangame.servletcontainerinit.SpringInitializer
+com.proxiad.task.ivanboyukliev.hangmangame.SpringInitializer
